[refactor] Mudança na estrutura do projeto. Nenhum código ou lógica alterada
