+++ src/main/java/com/prpa/bancodigital/repository/dao/AbstractDao.java
@@ -1,5 +1,6 @@
-package com.prpa.bancodigital.dao;
+package com.prpa.bancodigital.repository.dao;
 
+import com.prpa.bancodigital.config.respository.QueryResolver;
 import lombok.Getter;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageImpl;
