+++ src/main/java/com/prpa/bancodigital/repository/dao/Dao.java
@@ -1,4 +1,4 @@
-package com.prpa.bancodigital.dao;
+package com.prpa.bancodigital.repository.dao;
 
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
